gBenchmark: Remove Duplicate
Remove duplicate `gbenchmark` library
and keep its patch to remove the shipped
-Werror
